% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{estimate_horizon}
\alias{estimate_horizon}
\title{Estimate the horizon location}
\usage{
estimate_horizon(img, penalty = 0.05)
}
\arguments{
\item{img}{single raster layer to process}

\item{penalty}{parameter to set the sensitivity of the changepoint
detection used in finding the horizon}
}
\description{
Uses breakpoint detection on the blue channel to
detect the difference between sky and vegetation.
}
\examples{
# no examples yet
}
\keyword{interest}
\keyword{of}
\keyword{region}
\keyword{selection}
