% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{estimate_roi}
\alias{estimate_roi}
\title{Estimates a region of interest (ROI)}
\usage{
estimate_roi(img, padding = 0.1)
}
\arguments{
\item{img}{RGB image to process (filename or 3-layer RGB stack or brick)}

\item{padding}{percentage of the image width / height to pad}

\item{plot}{plot resulting image with all available information}
}
\description{
Uses the approximate horizon location in an image.
}
\keyword{interest}
\keyword{of}
\keyword{region}
\keyword{selection}
