% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_mapping_functions.R
\name{map_summary_values}
\alias{map_summary_values}
\title{Summary stats mapping function}
\usage{
map_summary_values(
  data,
  spatial_unit = "worldclim_2.5",
  value = "nr_fields",
  inset = TRUE,
  background,
  title = "title",
  subtitle = "subtitle",
  legend = "legend",
  filename
)
}
\arguments{
\item{data}{cgiar big data summary file, subset by season}

\item{spatial_unit}{which spatial unit to use (default = worldclim_2.5)}

\item{value}{which value to plot}

\item{inset}{plot an inset of India on the graph}

\item{background}{location of a background raster (DEM)}

\item{title}{plot title}

\item{subtitle}{plot subtitle}

\item{legend}{plot legend}

\item{filename}{filename where to export the resulting png}
}
\value{
a plot of pbi data
}
\description{
Summary stats mapping function
}
