% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{optimal_span}
\alias{optimal_span}
\title{Calculates the optimal span for a loess spline}
\usage{
optimal_span(
  y,
  x = NULL,
  weights = NULL,
  start = 0.01,
  step = 0.01,
  label = NULL,
  plot = FALSE
)
}
\arguments{
\item{y:}{a vector with measurement values to smooth}

\item{x:}{a vector with dates / time steps}

\item{weights:}{optional values to weigh the loess fit with}

\item{step:}{span increment size}
}
\description{
Uses an Akaike information criterion (AIC) to balance
function complexity and available data.
}
\examples{

\dontrun{
# Internal function only, should not be used stand-alone.

l = sin(1,10,0.01)
l = l + runif(length(l)))
optimal.span(l, plot = TRUE)
}
}
\keyword{loess,}
\keyword{series}
\keyword{smoother,}
\keyword{span,}
\keyword{time}
