% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_anonymize_data.R
\name{pbi_anonymize_data}
\alias{pbi_anonymize_data}
\title{Anonymize a raw data frame}
\usage{
pbi_anonymize_data(
  df,
  method = "sedac",
  country = "IND",
  resolution = 2.5,
  rds = "./data-raw/sedac_indian_villages.rds",
  output_file
)
}
\arguments{
\item{df}{a dataframe or filename of a data file}

\item{method}{the method to use (gadm or grid)}

\item{country}{country to use in the gadm method}

\item{resolution}{resolution to use in the grid method (2.5, 5 or 10)}

\item{rds}{rds file with sf polygons of villages}

\item{output_file}{where to store the results}
}
\value{
dataframe with lat lon values mapped to centroids
}
\description{
Anonymize raw data frame with colour indices
and other field / farmer specific data and
overwrite the original image coordinates with
those of an aggregated unit, either a WorldClim
grid cell or GADM administrative town boundaries.
}
