% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_colour_metrics.r
\name{pbi_colour_metrics}
\alias{pbi_colour_metrics}
\title{Estimates the gcc value for a region of interest}
\usage{
pbi_colour_metrics(img, roi)
}
\arguments{
\item{img}{RGB image to process (filename or 3-layer RGB stack or brick)}

\item{roi}{existing roi description image}

\item{plot}{plot resulting image with all available information}
}
\description{
Estimates the gcc value for a region of interest
}
\keyword{calculation}
\keyword{gcc}
