% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_map_coordinates.R
\name{pbi_map_sedac}
\alias{pbi_map_sedac}
\title{Maps locations to town centroids}
\usage{
pbi_map_sedac(df, rds)
}
\arguments{
\item{df}{data frame containing lat lon locations}

\item{rds}{rds file which includes an sf object with the village
parameters}
}
\value{
dataframe with lat lon values mapped to centroids
}
\description{
For a particular country maps locations of fields to the
centroid of towns. This anonymizes the data by obfuscating
the original location.
}
