% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_map_coordinates.R
\name{pbi_map_shp}
\alias{pbi_map_shp}
\title{Maps locations to arbitrary shape file centroids}
\usage{
pbi_map_shp(df, poly, project_id)
}
\arguments{
\item{df}{data frame containing lat lon locations}

\item{poly}{SpatialPolygon dataframe from which to take coordinates}

\item{project_id}{unique project id (numerical value)}
}
\value{
dataframe with lat lon values mapped to centroids
}
\description{
Maps locations to arbitrary shape file centroids
}
