% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_process_images.r
\name{pbi_process_images}
\alias{pbi_process_images}
\title{Batch process all IFPRI PBI data}
\usage{
pbi_process_images(
  database,
  img_path,
  output_path,
  output_prefix = "pbi_rabi_2016_2017",
  season_id,
  ncores = 12
)
}
\arguments{
\item{database}{RDS or DTA database file as provided by IFPRI}

\item{img_path}{local IFPRI database of images to process}

\item{output_path}{where to store processing results}

\item{output_prefix}{prefix to use on processed data}

\item{ncores}{numbers of cores to use in parallel processing}
}
\description{
Uses an RDS data file as a driver
}
\keyword{QA/GC}
\keyword{calculation,}
\keyword{gcc}
