% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{phenophases}
\alias{phenophases}
\title{Calculate phenophases from a smoothed curve}
\usage{
phenophases(date, value, span = 0.4, long_format = TRUE)
}
\arguments{
\item{date}{date vector}

\item{value}{gcc_90 value}

\item{span}{smoothing value to use when estimating phenophases}

\item{long_format}{}
}
\description{
Will return valid values if smoothing works,
will return NA if it fails.
}
\keyword{phenology,}
\keyword{phenophases}
