% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{read_size}
\alias{read_size}
\title{Reads and sizes images or raster layers
to a consistent format}
\usage{
read_size(img, ncols = 640, nrows = NULL)
}
\arguments{
\item{img}{RGB image, or raster file to process}

\item{ncols}{resize to a particular number of columns}

\item{nrows}{resize to a particular number of rows
if NULL the same ratio is used as set by the original ncol
vs. the specified ncols.}
}
\description{
Reads and sizes images or raster layers
to a consistent format
}
\keyword{image}
\keyword{preparation,}
\keyword{reading}
