% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{smooth_ts}
\alias{smooth_ts}
\title{Calculates a smoothed fit for a set of values, a span value
can be set / found by using optimal_span()}
\usage{
smooth_ts(x, y, span = 0.4)
}
\arguments{
\item{x}{a vector with dates / time steps}

\item{y}{a vector with measurement values to smooth}

\item{span}{optional values to weigh the loess fit with}
}
\description{
Calculates a smoothed fit for a set of values, a span value
can be set / found by using optimal_span()
}
\keyword{loess,}
\keyword{series}
\keyword{smoother,}
\keyword{span,}
\keyword{time}
